Functionality improvements in read/write options, menu features implemented

+++ FileReaderWriter/Menu/MenuStates/ReadMenuState.cs
@@ -1,4 +1,5 @@
 using System;
+using System.IO;
 using FileReaderWriter.ReadOptions;
 using FileReaderWriter.WriteOptions;
 
@@ -8,7 +9,24 @@ namespace FileReaderWriter.Menu.MenuStates
     {
         public override void DisplayMenu()
         {
-            throw new NotImplementedException();
+            Console.Clear();
+
+            Console.WriteLine("READ MENU\n" +
+                "—————————————————————————————\n" +
+                "1 -- Read from (only txt, rtxt, etxt, btxt)\n" +
+                "2 -- Back to the main menu");
+
+            switch (Console.ReadKey().Key)
+            {
+                case ConsoleKey.D1:
+                    Console.WriteLine("Specify path to the file..\n" +
+                        @"Example: C:\ExampleFolder\example.txt");
+                    _menuContext.ReadFromSpecificFile();
+                    break;
+                case ConsoleKey.D2:
+                    _menuContext.PressBack();
+                    break;
+            }
         }
 
         public void InformationAboutMenuState()
@@ -28,22 +46,30 @@ namespace FileReaderWriter.Menu.MenuStates
 
         public override void ReadFromSpecificFile()
         {
-            Console.WriteLine("You've chosen read from RTXT. Specify path to the file...");
-            string path = Console.ReadLine();
+            string? path = Console.ReadLine();
 
-            FileReader reader = new FileReader(new RTxtFormatter());
-            string content = reader.ReadContentFromFile(path);
-            string formattedContent = reader.FormatContent(content);
+            FileReader fileReader = new FileReader();
 
-            FileWriter writer = new FileWriter(new TxtWriter());
-            writer.WriteToFile(formattedContent, @"E:\Study\Test.txt");
+            if (File.Exists(path))
+            {
+                string content = fileReader.ReadContentFromFile(path);
 
+                FileWriter fileWriter = new FileWriter(new TxtWriter());
 
+                Console.WriteLine("Specify destination txt file...");
+
+                string? targetFile = Console.ReadLine();
+
+                if (File.Exists(targetFile))
+                {
+                    fileWriter.WriteToFile(content, targetFile);
+                }
+            }
         }
 
         public override void WriteToSpecificFile()
         {
-            throw new NotImplementedException();
+            Console.WriteLine("That's doing nothing here");
         }
     }
 }
